AdminVerifyMethod
It's a method to verify if a user is the admin of the post. The parameters are request parameters (postId, userId). This method is useful for the frontend so that it can determine whether the 'advertise' button should be displayed to the user or not, depending on whether the user is the admin of the post or not."